requirements.txt
aiogram>=3.4
fastapi>=0.110
uvicorn[standard]>=0.27
aiosqlite>=0.20
pydantic>=2.6
python-dotenv>=1.0
